pyproject.toml
[tool.poetry]
name = "github-automatically-create-jira-issues-when-advanced-scanning-detects-an-issue"
version = "0.1.0"
description = ""
authors = ["githubofkrishnadhas <krishnadhasnk1997@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
atlassian-python-api = "^3.41.13"
python-dotenv = "^1.0.1"
requests = "^2.32.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
